legal_system_project - Qwen/requirements.txt
dashscope
gradio
openai
easyocr
PyPDF2
python-docx
numpy
opencv-python
langchain
fastapi
uvicorn
sse_starlette
sentence_transformers
faiss-cpu
streamlit
chardet
tiktoken
websockets
python-multipart
torch
torchvision
torchaudio